Add files via upload
PHP-SMK Minggu ke 12 Project Restoran

+++ PHP-SMK/restoran/kategori/delete.php
@@ -1,10 +1,10 @@
 <?php
+if (isset($_GET['id'])) {
+    $id = $_GET['id'];
 
-require_once "../function.php";
+    $sql = "DELETE FROM  tblkategori  WHERE idkategori=$id";
 
+    $db->runSQL($sql);
 
-$sql = "DELETE FROM tblkategori WHERE idkategori = $id";
-
-$result = mysqli_query($koneksi, $sql);
-
-header("location:http://localhost/PHP-SMK/restoran/kategori/select.php");
+    header("location:?f=kategori&m=select");
+}
